SSE turned to class
